Remove red bubble data from window state
